+++ pymove/tests/test_base_init.py
@@ -1,5 +1,6 @@
 try:
     from pymove import *  # noqa
+
     _top_import_error = None
 except Exception as e:
     _top_import_error = e
